Issue #7689: Move Profiler interface from concept-engine to concept-base.
