+++ components/concept/base/src/main/java/mozilla/components/concept/base/profiler/Profiler.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.concept.engine.profiler
+package mozilla.components.concept.base.profiler
 
 /**
  * [Profiler] is being used to manage Firefox Profiler related features.
